requirements.txt
pyautogen==0.2.35
streamlit>=1.35.0
python-dotenv
